Fix incompatible to_string function declaration when size_t is unsigned int

+++ src/string_utils.hh
@@ -87,7 +87,7 @@ constexpr Hex hex(size_t val) { return {val}; }
 InplaceString<15> to_string(int val);
 InplaceString<15> to_string(unsigned val);
 InplaceString<23> to_string(long int val);
-InplaceString<23> to_string(size_t val);
+InplaceString<23> to_string(unsigned long val);
 InplaceString<23> to_string(long long int val);
 InplaceString<23> to_string(Hex val);
 InplaceString<23> to_string(float val);
